requirements.txt
flask==3.0.2
python-dotenv==1.0.1
oracledb==2.1.0